+++ src/basepkg/gp_decorators/__init__.py
@@ -1,1 +1,6 @@
-__author__ = 'sb'
+# coding=utf-8
+"""Decorators."""
+
+from .singleton import singleton
+
+__all__ = ['singleton']